recommend_agent/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
openai
requests
jinja2